requirements.txt
nltk==3.9.1
graphviz==0.20.3
pillow==11.1.0
